pyproject.toml
[project]
name = "hh-applications-bot"
version = "0.1.0"
description = "send automated job applications"
authors = [
    {name = "Dmitriy Korneev"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "hh-applications-bot"
packages = [
    { include = "src" }
]